+++ src/main/java/com/refinedmods/refinedstorage/mekanism/externalstorage/ChemicalExternalStorageProvider.java
@@ -1,9 +1,12 @@
-package com.refinedmods.refinedstorage.mekanism.importer;
+package com.refinedmods.refinedstorage.mekanism.externalstorage;
 
 import com.refinedmods.refinedstorage.api.core.Action;
-import com.refinedmods.refinedstorage.api.network.impl.node.importer.ImporterSource;
+import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.storage.Actor;
+import com.refinedmods.refinedstorage.api.storage.ExtractableStorage;
+import com.refinedmods.refinedstorage.api.storage.InsertableStorage;
+import com.refinedmods.refinedstorage.api.storage.external.ExternalStorageProvider;
 import com.refinedmods.refinedstorage.common.api.support.network.AmountOverride;
 import com.refinedmods.refinedstorage.mekanism.ChemicalCapabilityCache;
 import com.refinedmods.refinedstorage.mekanism.ChemicalExtractableStorage;
@@ -11,20 +14,20 @@ import com.refinedmods.refinedstorage.mekanism.ChemicalInsertableStorage;
 
 import java.util.Iterator;
 
-class ChemicalImporterSource implements ImporterSource {
+class ChemicalExternalStorageProvider implements ExternalStorageProvider {
     private final ChemicalCapabilityCache capabilityCache;
-    private final ChemicalInsertableStorage insertTarget;
-    private final ChemicalExtractableStorage extractTarget;
+    private final InsertableStorage insertTarget;
+    private final ExtractableStorage extractTarget;
 
-    ChemicalImporterSource(final ChemicalCapabilityCache capabilityCache, final AmountOverride amountOverride) {
+    ChemicalExternalStorageProvider(final ChemicalCapabilityCache capabilityCache) {
         this.capabilityCache = capabilityCache;
         this.insertTarget = new ChemicalInsertableStorage(capabilityCache, AmountOverride.NONE);
-        this.extractTarget = new ChemicalExtractableStorage(capabilityCache, amountOverride);
+        this.extractTarget = new ChemicalExtractableStorage(capabilityCache, AmountOverride.NONE);
     }
 
     @Override
-    public Iterator<ResourceKey> getResources() {
-        return capabilityCache.createIterator();
+    public Iterator<ResourceAmount> iterator() {
+        return capabilityCache.createAmountIterator();
     }
 
     @Override